+++ src/device/device.cpp
@@ -363,8 +363,13 @@ void Device::run()
 	// Init all remaining subsystems
 	_display = display::create(_allocator);
 
-	_width  = _boot_config.window_w;
-	_height = _boot_config.window_h;
+	if (_options._window_width.has_changed() || _options._window_height.has_changed()) {
+		_width  = _options._window_width;
+		_height = _options._window_height;
+	} else {
+		_width  = _boot_config.window_w;
+		_height = _boot_config.window_h;
+	}
 
 	_window = window::create(_allocator);
 	_window->open(_options._window_x